python/outdated-terraform-modules/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
entrypoints==0.3
flake8==3.7.8
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-gitlab==1.11.0
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
